feat: add page and table actions for v2 app
