+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -14,7 +14,8 @@ import { ToggleClearFilterBarQuickAction } from './lr-toggle-clear-filter-bar';
 import { ChangeTableColumnsQuickAction } from './change-table-columns';
 import { AddHeaderFieldQuickAction } from '../common/op-add-header-field';
 import { AddCustomSectionQuickAction } from '../common/op-add-custom-section';
-
+import { AddTableActionQuickAction } from '../fe-v2/create-table-action';
+import { AddPageActionQuickAction } from '../fe-v2/create-page-action';
 type PageName = 'listReport' | 'objectPage';
 
 const OBJECT_PAGE_TYPE = 'sap.suite.ui.generic.template.ObjectPage.view.Details';
@@ -40,7 +41,9 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                     definitions: [
                         ToggleClearFilterBarQuickAction,
                         AddControllerToPageQuickAction,
-                        ChangeTableColumnsQuickAction
+                        ChangeTableColumnsQuickAction,
+                        AddTableActionQuickAction,
+                        AddPageActionQuickAction
                     ],
                     view,
                     key: name + index
@@ -52,7 +55,9 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                         AddControllerToPageQuickAction,
                         ChangeTableColumnsQuickAction,
                         AddHeaderFieldQuickAction,
-                        AddCustomSectionQuickAction
+                        AddCustomSectionQuickAction,
+                        AddTableActionQuickAction,
+                        AddPageActionQuickAction
                     ],
                     view,
                     key: name + index
