+++ packages/preview-middleware-client/src/messagebundle.properties
@@ -2,6 +2,9 @@ QUICK_ACTION_ADD_PAGE_CONTROLLER=Add Controller to Page
 QUICK_ACTION_SHOW_PAGE_CONTROLLER=Show Page Controller
 QUICK_ACTION_OP_ADD_HEADER_FIELD=Add Header Field
 QUICK_ACTION_OP_ADD_CUSTOM_SECTION=Add Custom Section
+QUICK_ACTION_ADD_CUSTOM_PAGE_ACTION=Add Custom Page Action
+QUICK_ACTION_ADD_CUSTOM_TABLE_ACTION=Add Custom Table Action
+
 
 V2_QUICK_ACTION_CHANGE_TABLE_COLUMNS=Change Table Columns
 
@@ -10,6 +13,7 @@ V2_QUICK_ACTION_LR_DISABLE_CLEAR_FILTER_BAR=Disable "Clear" Button in Filter Bar
 
 
 V4_QUICK_ACTION_CHANGE_TABLE_COLUMNS=Change Table Columns
+V4_QUICK_ACTION_ADD_CUSTOM_TABLE_ACTION=Add Custom Table Action
 
 V4_QUICK_ACTION_LR_ENABLE_CLEAR_FILTER_BAR=Enable "Clear" Button in Filter Bar
 V4_QUICK_ACTION_LR_DISABLE_CLEAR_FILTER_BAR=Disable "Clear" Button in Filter Bar
